Fixed: Full screen startup
The monitor preference checks were not being performed - ever - when running experiments. I separated the check on exp settings and on monitor settings. Each is performed at an appropriate time during the process of running an experiment. The experiment check just checks ExperimentSettings, the monitor check looks at the current monitor assignments, but only cares if the expt is to be run at full screen.