Task: Full screen startup
Remove calls to showFullScreen(). Replace with simple call to show(). 
The showFullScreen call had different behavior on MacOS 10.10 (not sure about 10.11, 10.12) than on 10.13 (High Sierra). It seems that on the older OS the "screen" was the entire desktop, so "full screen" covered the entire desktop, and the ControlPanel dialog was relegated to an off-screen location. Something about how Mac uses "spaces" and such. I'm not going to try and understand, just fixing and moving on.