re: Accumulated looking should stop trial at threshold
When total looking time habituation is specified, implement an additional state for a trial end. The Looker will emit a signal "phaseAccumulatedLookTime()" when the total looking time for habituation is reached. The looking time at which to trigger this signal will be different each trial, depending on the amount of looking.