Re-create screenStarted signal.
This signal tells us when an individual stim screen has started. It
comes from gstreamer's state changes, as seen on the bus in messages
FROM one of the sinks. This particular signal (and the Event Log entry
that it led to) was lost in the gstreamer conversion. All this was
necessary to fix results output so it has the _actual_ filenames
available. I do it this way to ensure that something has actually
happened on that screen (or the speaker for sound sources), instead of
just generating these log events once the stim's main STATE_CHANGED msg
comes across the bus.

Closes: #3